// src/enums/messages.ts
export enum walletMessage {
    TRANSACTION_SUCCESS= 'پرداخت با موفقیت انجام شد',
    TRANSACTION_NOT_FOUND = 'تراکنش یافت نشد',
    TRANSACTION_DB_ERROR = 'خطا در ثبت تراکنش در دیتابیس',
    TRANSACTION_ERROR = 'پرداخت ناموفق بود',
    INSUFFICIENT_BALANCE_WALLET = 'موجودی کیف پول شما کافی نیست',
    WITHDRAW_SUCCESSFULL = 'برداشت با موفقیت انجام شد'
}

export enum userMessages {
    USER_NOT_FOUND = 'کاربر یافت نشد'
}